pyproject.toml
[tool.poetry]
name = "ssss"
version = "1.1.0"
description = "ssss is a static site generator that uses Jinja2 and Markdown to build fast, lightweight static HTML sites. Scaffold a new site in seconds with --scaffold and run it anywhere."
authors = ["Magnus Åberg <himself@magnusaberg.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
markdown = "^3.4.3"
staticjinja = "^4.1.3"
pyyaml = "^6.0.2"
argparse = "^1.4.0"
importlib-metadata = "^6.3.0"
jinja2 = "^3.1.6"
zipp = ">=3.19.1"

[tool.poetry.scripts]
ssss = "ssss:main"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^6.0.0"
commitizen = "^4.13.9"
pre-commit = "^4.5.1"
flake8 = "^7.3.0"

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "v$version"
version_scheme = "pep440"
version_provider = "poetry"
update_changelog_on_bump = true
changelog_file = "CHANGELOG.md"
major_version_zero = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
